Update tsconfig.json
The announcement of TypeScript 5.1 is here https://devblogs.microsoft.com/typescript/announcing-typescript-5-1/
There are some changes that are not mentioned in the release notes.

It is no longer possible to specify `help` of command line options in tsconfig.json.
In line with this, tsconfig.json is recreated with the latest version. Create it with `rm tsconfig.json; yarn tsc --init`.

Also, I confirmed that `yarn build` did not change the generated files by this.